requirements.txt
# Core Dependencies
opencv-python>=4.8.0
pytesseract>=0.3.10
Pillow>=10.0.0
numpy>=1.24.0

# NLP & PII Detection
spacy>=3.6.0

# Text Processing
textblob>=0.17.1

# Data Handling
pandas>=2.0.0

# Jupyter Notebook
jupyter>=1.0.0
ipykernel>=6.25.0
matplotlib>=3.7.0

# Utilities
tqdm>=4.65.0
